renaming from new to New (2 step)
renaming file (second step)


.
